# Conflicts:
#	folib-web-core/src/main/java/com/veadan/folib/controllers/BaseController.java
#	folib-web-core/src/main/java/com/veadan/folib/controllers/layout/huggingface/HuggingFaceMLControllers.java
#	folib-web-vue/src/views/Storage/Storages.vue

+++ folib-web-core/src/main/java/com/veadan/folib/components/webhook/BaseWebhookEventsProvider.java
@@ -124,8 +124,17 @@ public abstract class BaseWebhookEventsProvider implements WebhookEventsProvider
         String storageId = repositoryPath.getStorageId(), repositoryId = repositoryPath.getRepositoryId(), name = repositoryPath.getFileName().toString(), path = artifactData.getPath(), failureReason = "";
         try {
             if (Files.exists(repositoryPath)) {
-                RepositoryFiles.delete(repositoryPath);
-                result = checkNotExists(repositoryPath);
+                if (2 == type) {
+                    //定时任务触发，直接删除
+                    RepositoryFiles.delete(repositoryPath);
+                    result = checkNotExists(repositoryPath);
+                } else {
+                    //webhook触发，保存至数据库，异步定时任务处理
+                    WebhookEventsLog webhookEventsLog = WebhookEventsLog.builder().eventType(webhook.getEventType()).eventRepositoryId(artifactData.getRepoKey()).storageId(storageId).repositoryId(repositoryId).artifactName(name)
+                            .artifactPath(path).sha256Checksum(artifactData.getSha256()).size(artifactData.getSize()).status(WebhookEventsStatusEnum.INIT.getStatus()).build();
+                    webhookEventsLogService.saveWebhookEventsLog(webhookEventsLog, type);
+                    result = true;
+                }
             }
         } catch (Exception ex) {
             log.error("Webhook event handler eventRepositoryId [{}] storageId [{}] repositoryId [{}] path [{}] error [{}]", artifactData.getRepoKey(), storageId, repositoryId, path, ExceptionUtils.getStackTrace(ex));
